requirements.txt
# Core dependencies
kaggle==1.6.6
requests==2.31.0
python-dotenv==1.0.0

# Google Gemini AI
google-generativeai==0.3.2

# Data processing
pandas==2.1.4
numpy==1.26.2

# Web scraping
beautifulsoup4==4.12.2
selenium==4.16.0

# Research papers
arxiv==2.1.0

# Template engines
jinja2==3.1.2

# GitHub API
PyGithub==2.1.1

# Scheduling and async
schedule==1.2.0
aiohttp==3.9.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
